refactor: TIL 모아보기 페이지 컴포넌트 폴더구조 변경
