--- src/components/roadmap/manage/TIL/ManagePeopleTIL/style.ts
@@ -1,23 +1,0 @@
-import styled from '@emotion/styled';
-
-export const SearchSection = styled.section`
-  display: flex;
-  flex-direction: column;
-`;
-
-export const MainSearchContainer = styled.div`
-  display: flex;
-  gap: 2rem;
-`;
-
-export const Container = styled.div`
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  margin-top: 1rem;
-`;
-
-export const TILCount = styled.div`
-  font-weight: 600;
-  padding-left: 4px;
-`;
